requirements.txt
Flask
firebase-admin
stripe
python-dotenv
